feat: adds ability to detect Python virtual envs
