+++ package.json
@@ -761,6 +761,7 @@
     "@typescript-eslint/parser": "^8.24.0",
     "@vscode/test-electron": "^2.4.1",
     "c8": "^10.1.3",
+    "copy-webpack-plugin": "^12.0.2",
     "eslint": "^9.20.1",
     "eslint-plugin-import": "^2.29.1",
     "eslint-plugin-jsdoc": "^50.2.2",
