+++ src/util/tools.ts
@@ -1,11 +1,11 @@
 import * as assert from 'assert';
-import * as cp from 'child_process';
 import * as os from 'os';
 import * as path from 'path';
 
 import * as vscode from 'vscode';
 
 import { isString, isArrayOfString } from './helper';
+import { pipInstall } from './python';
 
 export const LS_NAME = 'fortls';
 export const EXTENSION_ID = 'fortran';
@@ -138,42 +138,6 @@ export async function promptForMissingTool(
 }
 
 /**
- * A wrapper around a call to `pip` for installing external tools.
- * Does not explicitly check if `pip` is installed.
- *
- * @param pyPackage name of python package in PyPi
- */
-export async function pipInstall(pyPackage: string): Promise<string> {
-  const py = 'python3'; // Fetches the top-most python in the Shell
-  const args = ['-m', 'pip', 'install', '--user', '--upgrade', pyPackage];
-  return await shellTask(py, args, `pip: ${pyPackage}`);
-}
-
-export async function shellTask(command: string, args: string[], name: string): Promise<string> {
-  const task = new vscode.Task(
-    { type: 'shell' },
-    vscode.TaskScope.Workspace,
-    name,
-    'Modern Fortran',
-    new vscode.ShellExecution(command, args)
-  );
-  // Temporay fix to https://github.com/microsoft/vscode/issues/157756
-  (<vscode.Task>task).definition = { type: 'shell', command: command };
-  const execution = await vscode.tasks.executeTask(task);
-  return await new Promise<string>((resolve, reject) => {
-    const disposable = vscode.tasks.onDidEndTaskProcess(e => {
-      if (e.execution === execution) {
-        disposable.dispose();
-        if (e.exitCode !== 0) {
-          reject(`ERROR: ${e.execution.task.name} failed with code ${e.exitCode}`);
-        }
-        resolve(`${name}: shell task completed successfully.`);
-      }
-    });
-  });
-}
-
-/**
  * Resolve VSCode internal variables `workspaceFolder`, `env`, `config`, etc.
  *
  * @param input string for which we will substitute internal variables
@@ -285,48 +249,4 @@ export function pathRelToAbs(relPath: string, uri: vscode.Uri): string | undefin
 
 export function getWholeFileRange(document: vscode.TextDocument): vscode.Range {
   return new vscode.Range(0, 0, document.lineCount, 0);
-}
-
-/**
- * Spawn a command as a `Promise`
- * @param cmd command to execute
- * @param args arguments to pass to the command
- * @param options child_process.spawn options
- * @param input any input to pass to stdin
- * @param ignoreExitCode ignore the exit code of the process and `resolve` the promise
- * @returns Tuple[string, string] `[stdout, stderr]`. By default will `reject` if exit code is non-zero.
- */
-export async function spawnAsPromise(
-  cmd: string,
-  args: ReadonlyArray<string> | undefined,
-  options?: cp.SpawnOptions | undefined,
-  input?: string | undefined,
-  ignoreExitCode?: boolean
-) {
-  return new Promise<[string, string]>((resolve, reject) => {
-    let stdout = '';
-    let stderr = '';
-    const child = cp.spawn(cmd, args, options);
-    child.stdout.on('data', data => {
-      stdout += data;
-    });
-    child.stderr.on('data', data => {
-      stderr += data;
-    });
-    child.on('close', code => {
-      if (ignoreExitCode || code === 0) {
-        resolve([stdout, stderr]);
-      } else {
-        reject([stdout, stderr]);
-      }
-    });
-    child.on('error', err => {
-      reject(err.toString());
-    });
-
-    if (input) {
-      child.stdin.write(input);
-      child.stdin.end();
-    }
-  });
 }
